Remove dashboard layout abstractions
