--- web-common/src/features/dashboards/workspace/DashboardControls.svelte
@@ -1,13 +1,0 @@
-<script lang="ts">
-  import Filters from "../filters/Filters.svelte";
-  import TimeControls from "../time-controls/TimeControls.svelte";
-
-  export let metricViewName: string;
-</script>
-
-<div class="p-1 py-2 space-y-2">
-  <TimeControls {metricViewName} />
-  {#key metricViewName}
-    <Filters />
-  {/key}
-</div>
